requirements.txt
fastapi[standard]
uvicorn
pydantic
pydantic_settings
requests
httpx
python-dotenv
sse-starlette
pyyaml